requirements.txt
# Core
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# LangChain
langchain>=0.1.0
langchain-classic>=1.0.0
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-huggingface>=0.0.1
langchain-community>=0.0.13

# Vector Store
faiss-cpu>=1.7.4

# OpenAI
openai>=1.0.0
langchain-openai>=0.0.5

# Embeddings (HuggingFace)
sentence-transformers>=2.2.0

# Document Processing
pdfplumber>=0.10.0
python-docx>=1.1.0
python-pptx>=0.6.23
pillow>=10.0.0

# OCR (optional)
pytesseract>=0.3.10

# File handling
filelock>=3.13.0

# Streamlit frontend
streamlit>=1.29.0
requests>=2.31.0
